+++ RELEASES.md
@@ -1209,3 +1209,13 @@ Based on:
 - [typescript v0.6.0-alpha.28] .
 ### Releases
 - [NPM v0.6.0-alpha.28] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.28 - .
+
+## 2026-03-02 08:13:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.736.3 (2.845.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0-alpha.29] .
+### Releases
+- [NPM v0.6.0-alpha.29] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.29 - .